Add a `Client` implementation for Elasticsearch 6
Add an implementation of the `Client` class which wraps an Elasticsearch
connection and index for Elasticsearch 6, and expose it on the request
as `request.es6`.

This will enable existing code for setting up the Elasticsearch index
and reindexing annotations to be used with the ES6 cluster.

The `Client` classes now expose a `using_es6` property to enable
switching on the ES version in the small number of places where this is
necessary.